refactor_POSSIBLE_BREAK_(reordered directory structure and merged components)
