+++ client/src/components/index.js
@@ -1,0 +1,5 @@
+import Navbar from './Navbar';
+import Footer from './Footer';
+import Facebook from './Facebook';
+
+export { Navbar, Footer, Facebook };
