+++ client/src/constants.js
@@ -1,6 +1,17 @@
 /**
- * Server API Paths
+ * Setup
+ */
+export const CURRENCY_PRE = '$'; // $ 100
+export const CURRENCY_POST = ''; // 100 $
+
+/**
+ * Login
  */
 export const API_PATH_LOGIN = '/api/auth/login';
 export const API_PATH_REGISTER = '/api/auth/register';
 export const API_PATH_FACEBOOK = '/api/auth/facebook/login';
+
+/**
+ * Items
+ */
+export const API_PATH_ITEMS = '/api/items';
