+++ client/src/routes/Home.spec.js
@@ -2,11 +2,12 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import { render } from '@testing-library/react';
 import axios from 'axios';
-import ItemDisplayComp from './ItemDisplayComp';
+import Home from './Home';
+import { API_PATH_ITEMS } from '../constants';
 
 jest.mock('axios');
 
-describe('<ItemDisplayComp />', () => {
+describe('<Home />', () => {
 	let data;
 
 	global.console = {
@@ -29,9 +30,9 @@ describe('<ItemDisplayComp />', () => {
 
 	it('should render', () => {
 		const spy = jest
-			.spyOn(ItemDisplayComp.prototype, 'componentDidMount')
+			.spyOn(Home.prototype, 'componentDidMount')
 			.mockImplementationOnce(() => data);
-		const wrapper = shallow(<ItemDisplayComp />);
+		const wrapper = shallow(<Home />);
 		wrapper.setState({ items: data });
 		expect(spy).toHaveBeenCalledTimes(1);
 		expect(wrapper.state().items).toHaveLength(1);
@@ -41,16 +42,16 @@ describe('<ItemDisplayComp />', () => {
 	it('should set item', async () => {
 		const resolve = Promise.resolve({ data });
 		axios.get.mockImplementationOnce(() => resolve);
-		const wrapper = shallow(<ItemDisplayComp />);
+		const wrapper = shallow(<Home />);
 		await resolve;
-		expect(axios.get).toHaveBeenCalledWith('api/items');
+		expect(axios.get).toHaveBeenCalledWith(API_PATH_ITEMS);
 		expect(wrapper.state().items).toMatchObject(data);
 	});
 	it('should log error', async () => {
 		try {
 			const reject = Promise.reject('error1');
 			axios.get.mockImplementationOnce(() => reject);
-			render(<ItemDisplayComp />);
+			render(<Home />);
 		} catch (e) {
 			expect(global.console.log).toHaveBeenCalledWith('error1');
 		}
