+++ client/src/setupTests.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import { configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
