+++ client/src/utils/index.js
@@ -1,0 +1,3 @@
+import { signOut, checkLoggedIn } from './LogInHandler';
+
+export { signOut, checkLoggedIn };
